requirements.txt
langchain-core>=0.2.0
langgraph>=0.0.17
langserve>=0.0.45
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
gigachat>=1.2.0